mchs/requirements.txt
aiofiles==0.8.0
aioredis==2.0.1
anyio==3.6.1
asks==3.0.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
blinker==1.4
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
priority==2.0.0
pyparsing==3.0.9
python-dotenv==0.20.0
quart==0.17.0
quart-trio==0.9.1
redis==4.3.3
requests==2.28.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.21.0
trio-asyncio==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
